Remove BronyTunes and EqBeats hourly regular sync
EqBeats is no longer actively in service, so there will be no new changes to the music data already present in the PVL database for it. BronyTunes was largely only included to pull items from the entire MLPMA (which they contained), but with the MLPMA being folded into Pony.fm, it can be used as the sole source for external music records.